Extract interfaces from Grid
Closes vaadin/framework8-issues#566
