+++ server/src/test/java/com/vaadin/tests/components/grid/GridNoSelectionModelTest.java
@@ -10,8 +10,8 @@ import org.junit.Test;
 
 import com.vaadin.server.data.provider.bov.Person;
 import com.vaadin.ui.Grid;
-import com.vaadin.ui.Grid.GridSelectionModel;
 import com.vaadin.ui.Grid.SelectionMode;
+import com.vaadin.ui.components.grid.GridSelectionModel;
 
 public class GridNoSelectionModelTest {
