+++ server/src/test/java/com/vaadin/tests/server/component/grid/GridDefaultHeaderTest.java
@@ -25,7 +25,7 @@ import org.junit.Test;
 import com.vaadin.data.ValueProvider;
 import com.vaadin.ui.Grid;
 import com.vaadin.ui.Grid.Column;
-import com.vaadin.ui.Grid.HeaderRow;
+import com.vaadin.ui.components.grid.HeaderRow;
 
 public class GridDefaultHeaderTest {
     private Grid<String> grid;
